stores owning, rights for deleting and editing stores

+++ internal/data/repository/store_repo_impl.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"marketplace/internal/domain/entities"
 	repository2 "marketplace/internal/domain/repository"
+	"marketplace/pkg/utils"
 	"sync"
 	"time"
 )
@@ -20,25 +21,22 @@ func NewStoreRepository(db *sql.DB) repository2.StoreRepository {
 	}
 }
 
-func (r *storeRepositoryImpl) Save(store entities.Store) error {
+func (r *storeRepositoryImpl) Save(store entities.Store, userID int64) error {
 	r.mu.Lock()
 	defer r.mu.Unlock()
-
-	var existingStoreID uint64
-	err := r.db.QueryRow("SELECT id FROM stores WHERE id = $1", store.ID).Scan(&existingStoreID)
-	if err == nil {
-		return errors.New("store already exists in the database")
-	}
-	if err != nil && err != sql.ErrNoRows {
-		return err
-	}
 
 	store.CreatedAt = time.Now()
 	store.UpdatedAt = store.CreatedAt
 
-	_, err = r.db.Exec(`INSERT INTO stores (name, description, owner_id, created_at, updated_at) 
-                        VALUES ($1, $2, $3, $4, $5)`,
-		store.Name, store.Description, store.OwnerID, store.CreatedAt, store.UpdatedAt)
+	var newStoreID uint64
+	err := r.db.QueryRow(`INSERT INTO stores (name, description, owner_id, created_at, updated_at) 
+                         VALUES ($1, $2, $3, $4, $5) RETURNING id`,
+		store.Name, store.Description, userID, store.CreatedAt, store.UpdatedAt).Scan(&newStoreID)
+	if err != nil {
+		return err
+	}
+
+	_, err = r.db.Exec(`UPDATE users SET owning_stores = array_append(owning_stores, $1) WHERE id = $2`, newStoreID, userID)
 	if err != nil {
 		return err
 	}
@@ -46,7 +44,7 @@ func (r *storeRepositoryImpl) Save(store entities.Store) error {
 	return nil
 }
 
-func (r *storeRepositoryImpl) FindByID(id uint64) (entities.Store, error) {
+func (r *storeRepositoryImpl) FindByID(id int64) (entities.Store, error) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 
@@ -65,25 +63,23 @@ func (r *storeRepositoryImpl) FindByID(id uint64) (entities.Store, error) {
 	return store, nil
 }
 
-func (r *storeRepositoryImpl) Update(store entities.Store) error {
+func (r *storeRepositoryImpl) Update(store entities.Store, userID int64) error {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 
-	var existingStoreID uint64
-	err := r.db.QueryRow("SELECT id FROM stores WHERE id = $1", store.ID).Scan(&existingStoreID)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return errors.New("store not found")
-		}
-		return err
+	storeExists, err := utils.CheckStoreExists(r.db, store.ID)
+	if err != nil || !storeExists {
+		return errors.New("store not found")
+	}
+
+	isOwner, err := utils.CheckUserOwnsStore(r.db, userID, store.ID)
+	if err != nil || !isOwner {
+		return errors.New("user does not own this store")
 	}
 
 	store.UpdatedAt = time.Now()
-
-	_, err = r.db.Exec(`UPDATE stores 
-                        SET name = $1, description = $2, owner_id = $3, updated_at = $4 
-                        WHERE id = $5`,
-		store.Name, store.Description, store.OwnerID, store.UpdatedAt, store.ID)
+	_, err = r.db.Exec(`UPDATE stores SET name = $1, description = $2, updated_at = $3 WHERE id = $4`,
+		store.Name, store.Description, store.UpdatedAt, store.ID)
 	if err != nil {
 		return err
 	}
@@ -91,17 +87,18 @@ func (r *storeRepositoryImpl) Update(store entities.Store) error {
 	return nil
 }
 
-func (r *storeRepositoryImpl) Delete(id uint64) error {
+func (r *storeRepositoryImpl) Delete(id int64, uid int64) error {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 
-	var existingStoreID uint64
-	err := r.db.QueryRow("SELECT id FROM stores WHERE id = $1", id).Scan(&existingStoreID)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return errors.New("store not found")
-		}
-		return err
+	storeExists, err := utils.CheckStoreExists(r.db, id)
+	if err != nil || !storeExists {
+		return errors.New("store not found")
+	}
+
+	isOwner, err := utils.CheckUserOwnsStore(r.db, uid, id)
+	if err != nil || !isOwner {
+		return errors.New("user does not own this store")
 	}
 
 	_, err = r.db.Exec("DELETE FROM stores WHERE id = $1", id)
